Inherit from EventEmitter
For the Clinic data server, I'm looking to mock `NOTIFY` queries/'notification' events.

The way I'm doing it is by expect()ing a NOTIFY and then emitting a
'notification' event by hand in the test. That requires that clients can
emit events, though :)

This seems more in the minimalist spirit of this module than parsing
NOTIFY queries and doing the emit in here.